+++ spec/lips.spec.js
@@ -397,6 +397,7 @@ describe('evaluate', function() {
         });
         it('should throw exception', function() {
             return lips.exec(code, env, env).catch(e => {
+                expect(e.code).toEqual('(apply f args)');
                 expect(e).toEqual(new Error("Unbound variable `f'"));
             });
         });